cleanup some code

+++ app/src/main/java/hr/bpervan/novaeva/fragments/EvaContentFragment.kt
@@ -231,7 +231,7 @@ class EvaContentFragment : EvaBaseFragment() {
         vijestWebView.loadHtmlText(evaContent?.text)
 
         /** Is this 'Duhovni poziv' or 'Odgovori' category?  */
-        if (categoryId == EvaCategory.VOCATION.id.toLong()) {
+        if (categoryId == EvaCategory.VOCATION.id) {
             btnPoziv.isVisible = true
             btnPoziv.setOnClickListener {
                 val text = "Hvaljen Isus i Marija, javljam Vam se jer razmišljam o duhovnom pozivu."
